tend: add tabbed native iphone satsang app
