+++ docs/coherence-substrate/satsang-guidance-event.form
@@ -7,18 +7,20 @@
 ; presence and turn mode before any transmission is asked for.
 ;
 ; Carrier:
-;   experiments/satsang-mac-app - SwiftUI Mac desktop app. It tails the local
-;   transcript JSONL, listens to the room microphone after explicit Start
-;   Listening, lets the holder edit utterances, and writes the request to
-;   ~/.coherence-network/satsang-guidance/events.jsonl plus the latest JSON and
-;   .form receipt. No external receiver is invoked by the GUI. Live room capture
-;   is the primary stream; Speech Recognition is a concurrent transcription side
-;   channel fed during that capture, not a before/after pass over stored audio.
-;   No-speech intervals restart only the recognition side channel and the visible
-;   mic level shows whether the room is reaching the input device. The mic meter
-;   comes alive separately from Speech Recognition authorization so permission
-;   delay is visible instead of silent. On Send, the GUI asks the local Form CLI
-;   first and records the
+;   experiments/satsang-mac-app - shared SwiftUI native app. macOS and iPhone
+;   entry points share one tabbed body: Room, Guidance, Memory, Learning,
+;   Resources, and Settings. The app tails the local transcript JSONL, listens to
+;   the room microphone after explicit Start Listening, lets the holder edit
+;   utterances, and writes the request to ~/.coherence-network/satsang-guidance/
+;   events.jsonl plus the latest JSON and .form receipt. No external receiver is
+;   invoked by the GUI. Live room capture is the primary stream; Speech
+;   Recognition is a concurrent transcription side channel fed during that
+;   capture, not a before/after pass over stored audio. No-speech intervals
+;   restart only the recognition side channel and the visible mic level shows
+;   whether the room is reaching the input device. The mic meter comes alive
+;   separately from Speech Recognition authorization so permission delay is
+;   visible instead of silent. On Send, the GUI asks the local Form CLI first and
+;   records the
 ;   body/RAG sufficiency receipt in the JSON and .form event; a remote oracle is
 ;   only requested as an explicit insufficient-local route, not called by the GUI.
 ;   The Send path also writes a local trusted room-memory receipt under
@@ -29,11 +31,13 @@
 ;   stream; macOS biometric speaker identification is outside the carrier.
 ;   File, process, audio-input, and speech-transcript resources cross one
 ;   generic host OS resource-door receipt in the Swift core. The receipt carries
-;   resolved macOS, Windows, and Android carrier mappings for every resource
-;   kind; live state is detected on macOS today, while Windows/Android device
-;   packages still promote declared doors to observed/open runtime receipts. The
-;   shared logic remains Form-native and rejects Python, Go, Rust, and
-;   TypeScript as app-boundary runtimes.
+;   resolved macOS, Windows, Android, and iPhone/iOS carrier mappings for every
+;   resource kind; live state is detected on Apple hosts today, while
+;   Windows/Android device packages still promote declared doors to observed/open
+;   runtime receipts. iOS cannot spawn arbitrary subprocesses, so its
+;   process-stdin/stdout door names an embedded Form runtime adapter until fkwu
+;   is packaged in-process. The shared logic remains Form-native and rejects
+;   Python, Go, Rust, and TypeScript as app-boundary runtimes.
 ;
 ; Body:
 ;   form/form-stdlib/satsang-guidance-event.fk decides whether the event is a
@@ -48,7 +52,7 @@
 ;   form/form-stdlib/satsang-host-boundary.fk names the generic host ABI,
 ;   detected resource doors, resolved platform carriers, and forbidden
 ;   app-boundary runtimes. Proof:
-;   form/form-stdlib/tests/satsang-host-boundary-band.fk -> 131071.
+;   form/form-stdlib/tests/satsang-host-boundary-band.fk -> 1048575.
 ;   form/form-stdlib/satsang-room-memory.fk names explicit local session memory,
 ;   speaker continuity, no-hidden-capture, and later context. Proof:
 ;   form/form-stdlib/tests/satsang-room-memory-band.fk -> 255.
