+++ experiments/satsang-mac-app/README.md
@@ -1,12 +1,14 @@
-# Satsang Mac App
+# Satsang Native App
 
-SwiftUI desktop carrier for a satsang guidance session.
+SwiftUI native carrier for a satsang guidance session.
 
-It listens to the room microphone after `Start Listening` is pressed, keeps that
-live capture as the primary stream, turns concurrent side-channel speech
-transcription into editable transcript lines, can also follow a local transcript
-file, shows every line that will be sent, and writes a
-`satsang-guidance-request` event for Sema or another invoked presence.
+The macOS app and iPhone app share one tabbed SwiftUI body: Room, Guidance,
+Memory, Learning, Resources, and Settings. The Room mode listens to the room
+microphone after `Start Listening` is pressed, keeps that live capture as the
+primary stream, turns concurrent side-channel speech transcription into editable
+transcript lines, can also follow a local transcript file, shows every line that
+will be sent, and writes a `satsang-guidance-request` event for Sema or another
+invoked presence.
 
 Default transcript:
 
@@ -35,6 +37,13 @@ Run from the repo root:
 swift run --package-path experiments/satsang-mac-app SatsangGuidance
 ```
 
+Run the iPhone SwiftUI shell as a package product on the host for source-level
+validation:
+
+```zsh
+swift run --package-path experiments/satsang-mac-app SatsangGuidancePhone
+```
+
 Build an `.app` bundle:
 
 ```zsh
@@ -59,13 +68,20 @@ sidecar must share the already-open listening stream. The memory proof lives in
 `form/form-stdlib/satsang-room-memory.fk`.
 
 The app boundary is intentionally small. Shared routing and sufficiency logic is
-Form-native; Swift is the current macOS host carrier for GUI, microphone, speech,
+Form-native; Swift is the current Apple host carrier for GUI, microphone, speech,
 file, and process resources. The request receipt names a generic host OS
-resource interface with resolved macOS, Windows, and Android carrier mappings.
+resource interface with resolved macOS, Windows, Android, and iPhone/iOS carrier
+mappings.
 Each Send receipt includes detected host resource doors for file read, file
 append, atomic file write, process stdin/stdout, audio input, and speech
 transcription, plus the cross-platform carrier matrix for those same doors.
 Python, Go, Rust, and TypeScript are not app-boundary runtimes for this carrier.
+Windows and Android are carrier mappings in this package, not full GUI apps yet.
+The iPhone target is native SwiftUI source; a device-signed archive still needs
+an Apple team/profile plus installed iOS SDK support. The permission metadata
+template lives at `Support/SatsangGuidancePhone/Info.plist`, and iOS routes the
+Form process door through an embedded runtime adapter rather than arbitrary
+subprocess spawning.
 
 Before Send writes the event, the app asks the repo-local native `form-cli`
 for a local body/RAG answer. The resulting route receipt is stored inside the
@@ -73,13 +89,13 @@ JSON and `.form` request. `remoteOracleRequested` is set only when the
 Form-native body and local RAG/local-LLM lane are not sufficient; the GUI does
 not call the remote oracle itself.
 
-The first use of `Start Listening` asks macOS for microphone and speech
+The first use of `Start Listening` asks the Apple host for microphone and speech
 recognition permission. The microphone meter starts as soon as microphone access
 is available. Speech Recognition then attaches as a side channel fed by the
 already-open live capture tap; it is not a before-recording or after-recording
 pass over stored audio. Partial speech appears in the transcript list as `room
-mic`; pressing `Stop Listening` commits the current partial line. If macOS
+mic`; pressing `Stop Listening` commits the current partial line. If Speech
 reports a no-speech interval, the listener keeps the capture stream open and
-restarts only the recognition side channel. The header shows a live microphone
+restarts only the recognition side channel. The Room tab shows a live microphone
 level while listening, so the holder can see whether the room is reaching the
 mic even while Speech permission is still pending.
