+++ experiments/satsang-mac-app/Sources/SatsangGuidancePhone/SatsangGuidancePhoneApp.swift
@@ -1,0 +1,11 @@
+import SatsangGuidanceKit
+import SwiftUI
+
+@main
+struct SatsangGuidancePhoneApp: App {
+    var body: some Scene {
+        WindowGroup {
+            SatsangGuidanceRootView(shell: .iPhone)
+        }
+    }
+}
