+++ experiments/satsang-mac-app/Support/SatsangGuidancePhone/Info.plist
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CFBundleDisplayName</key>
+	<string>Satsang Guidance</string>
+	<key>NSMicrophoneUsageDescription</key>
+	<string>Room listening uses the microphone after Start Listening is pressed.</string>
+	<key>NSSpeechRecognitionUsageDescription</key>
+	<string>Speech recognition turns live room audio into editable transcript lines.</string>
+</dict>
+</plist>
