+++ experiments/satsang-mac-app/Tests/SatsangMacCoreTests/SatsangMacCoreTests.swift
@@ -67,9 +67,10 @@ final class SatsangMacCoreTests: XCTestCase {
         XCTAssertTrue(form.contains("(host-resource-interface \"host-os-generic-resource-interface\")"))
         XCTAssertTrue(form.contains("(host-resource-door-count 6)"))
         XCTAssertTrue(form.contains("(host-resource-door-summary \"audio-input:declared:host-os-generic-resource-interface"))
-        XCTAssertTrue(form.contains("(host-platform-carrier-count 3)"))
+        XCTAssertTrue(form.contains("(host-platform-carrier-count 4)"))
         XCTAssertTrue(form.contains("windows:windows-minimal-host-carrier"))
         XCTAssertTrue(form.contains("android:android-minimal-host-carrier"))
+        XCTAssertTrue(form.contains("ios:iphone-minimal-host-carrier"))
         XCTAssertTrue(form.contains("(forbidden-runtime-carriers \"python,go,rust,typescript\")"))
         XCTAssertTrue(form.contains("(remote-oracle-requested 1)"))
         XCTAssertTrue(form.contains("(trusted-room-memory-context"))
@@ -128,22 +129,26 @@ final class SatsangMacCoreTests: XCTestCase {
         XCTAssertEqual(boundary.appBoundaryRuntimes, ["form", "swift-minimal-host-carrier"])
         XCTAssertEqual(boundary.forbiddenRuntimeCarriers, ["python", "go", "rust", "typescript"])
         XCTAssertEqual(boundary.resourceDoors.map(\.kind), boundary.allowedResourceKinds)
-        XCTAssertEqual(boundary.platformCarriers.map(\.platform), ["macos", "windows", "android"])
+        XCTAssertEqual(boundary.platformCarriers.map(\.platform), ["macos", "windows", "android", "ios"])
         XCTAssertTrue(boundary.platformCarriers.allSatisfy { $0.resourceDoors.count == boundary.allowedResourceKinds.count })
         XCTAssertTrue(boundary.platformTargets.contains("macos"))
         XCTAssertTrue(boundary.platformTargets.contains("windows"))
         XCTAssertTrue(boundary.platformTargets.contains("android"))
+        XCTAssertTrue(boundary.platformTargets.contains("ios"))
     }
 
-    func testWindowsAndAndroidHostCarriersResolveEveryDoor() {
+    func testWindowsAndroidAndIPhoneHostCarriersResolveEveryDoor() {
         let boundary = FormHostBoundaryReceipt()
         let windows = boundary.platformCarriers.first { $0.platform == "windows" }
         let android = boundary.platformCarriers.first { $0.platform == "android" }
+        let ios = boundary.platformCarriers.first { $0.platform == "ios" }
 
         XCTAssertEqual(windows?.hostCarrier, "windows-minimal-host-carrier")
         XCTAssertEqual(android?.hostCarrier, "android-minimal-host-carrier")
+        XCTAssertEqual(ios?.hostCarrier, "iphone-minimal-host-carrier")
         XCTAssertEqual(windows?.resourceDoors.map(\.kind), boundary.allowedResourceKinds)
         XCTAssertEqual(android?.resourceDoors.map(\.kind), boundary.allowedResourceKinds)
+        XCTAssertEqual(ios?.resourceDoors.map(\.kind), boundary.allowedResourceKinds)
         XCTAssertEqual(
             windows?.resourceDoors.first { $0.kind == "process-stdin-stdout" }?.carrier,
             "windows-createprocess-stdin-stdout"
@@ -152,8 +157,27 @@ final class SatsangMacCoreTests: XCTestCase {
             android?.resourceDoors.first { $0.kind == "speech-transcript" }?.carrier,
             "android-speechrecognizer"
         )
+        XCTAssertEqual(
+            ios?.resourceDoors.first { $0.kind == "speech-transcript" }?.carrier,
+            "ios-speech"
+        )
+        XCTAssertEqual(
+            ios?.resourceDoors.first { $0.kind == "process-stdin-stdout" }?.carrier,
+            "ios-embedded-form-cli-adapter"
+        )
         XCTAssertTrue(boundary.platformCarrierSummary.contains("android-audiorecord"))
+        XCTAssertTrue(boundary.platformCarrierSummary.contains("ios-avfoundation"))
         XCTAssertTrue(boundary.usesOnlyAllowedAppRuntimes)
+    }
+
+    func testNativeAppModesDefineSingleTabbedShell() {
+        let receipt = SatsangNativeAppModeReceipt()
+
+        XCTAssertEqual(receipt.kind, "satsang-native-tabbed-app")
+        XCTAssertEqual(receipt.singleAppBody, "form-native-satsang-guidance-body")
+        XCTAssertEqual(receipt.defaultMode, "room")
+        XCTAssertEqual(receipt.modes, ["room", "guidance", "memory", "learning", "resources", "settings"])
+        XCTAssertEqual(SatsangNativeAppMode.allCases.map(\.title), ["Room", "Guidance", "Memory", "Learning", "Resources", "Settings"])
     }
 
     func testTrustedRoomMemoryStoresSessionAndFeedsLaterContext() throws {
