+++ form/form-stdlib/tests/satsang-host-boundary-band.fk
@@ -3,7 +3,7 @@
 ; satsang-host-boundary-band — shared Form body, minimal host ABI, no app
 ; boundary Python/Go/Rust/TypeScript.
 ;
-; Verdict 131071:
+; Verdict 1048575:
 ;   1   Form is an allowed app runtime
 ;   2   Swift is only the minimal host carrier
 ;   4   Python is forbidden at the app boundary
@@ -14,6 +14,7 @@
 ;   128 macOS is a target platform
 ;   256 Windows is a target platform
 ;   512 Android is a target platform
+;   131072 iPhone/iOS is a target platform
 ;   1024 detected audio-input door is valid
 ;   2048 unknown resource kind is rejected as a detected door
 ;   4096 macOS carrier matrix resolves every resource door
@@ -21,6 +22,8 @@
 ;   16384 Android carrier matrix resolves every resource door
 ;   32768 Android speech carrier is the native speech recognizer door
 ;   65536 Windows process door is CreateProcess stdin/stdout
+;   262144 iPhone/iOS carrier matrix resolves every resource door
+;   524288 iPhone speech carrier is the native Speech side channel
 (do
     (let rec (shb-receipt "swift-minimal-host-carrier" "process-stdin-stdout" "macos"))
     (let audio-door (shb-door-receipt "audio-input" "open" "macos-avfoundation"))
@@ -46,5 +49,8 @@
     (let c14 (if (eq (shb-platform-carriers-complete? "android") 1) 16384 0))
     (let c15 (if (str_eq (shb-carrier-for "android" "speech-transcript") "android-speechrecognizer") 32768 0))
     (let c16 (if (str_eq (shb-carrier-for "windows" "process-stdin-stdout") "windows-createprocess-stdin-stdout") 65536 0))
+    (let c17 (if (eq (shb-platform-target? "ios") 1) 131072 0))
+    (let c18 (if (eq (shb-platform-carriers-complete? "ios") 1) 262144 0))
+    (let c19 (if (str_eq (shb-carrier-for "ios" "speech-transcript") "ios-speech") 524288 0))
 
-    (sum (list c0 c1 c2 c3 c4 c5 c6 c7 c8 c9 c10 c11 c12 c13 c14 c15 c16)))
+    (sum (list c0 c1 c2 c3 c4 c5 c6 c7 c8 c9 c10 c11 c12 c13 c14 c15 c16 c17 c18 c19)))
